requirements.txt
discord.py==1.3.3
tinydb==4.1.1
dateparser==0.7.6
python-dateutil==2.8.1